The next generation of tool: -reworked API for internal usage; -reworked interfaces to providers  conception for convinient usage; -added multiple group support; -added tooltip support for bricks;

+++ bundles/org.eclipse.chronograph.base/src/org/eclipse/chronograph/internal/base/UnitConverter.java
@@ -34,7 +34,7 @@ public class UnitConverter {
 		return (int) ChronoUnit.DAYS.between(HISTORICAL_TIME, targetDate);
 	}
 
-	public static LocalDate unitsToLocalDate(int units) {
+	public static LocalDate unitsToLocalDate(long units) {
 		return ChronoUnit.DAYS.addTo(HISTORICAL_TIME, units);
 	}
 }
